Refactoring, allow using default SAXParserFactory
