+++ src/main/java/openfasttrack/core/xml/event/Attribute.java
@@ -1,4 +1,4 @@
-package openfasttrack.core.xml;
+package openfasttrack.core.xml.event;
 
 /*-
  * #%L
